Fix Int to Int64 type conversion in DownloadManager
The totalFileSize parameter in DownloadedPlaylistInfo expects Int64,
but the sum query result was inferred as Int.

+++ mixbridge/Services/DownloadManager.swift
@@ -550,7 +550,7 @@ final class DownloadManager: ObservableObject {
                         soundCloudPlaylist: playlist.soundCloudPlaylist,
                         trackCount: playlistTrackIds.count,
                         downloadedTrackCount: downloadedCount,
-                        totalFileSize: totalSize
+                        totalFileSize: Int64(totalSize)
                     )
                 }
             }
